tests pass
TODO

Rename and rehierarch inlining
See if we can remove VarArray distinction / or add FixArray distinction
