+++ squeal-postgresql/squeal-postgresql.cabal
@@ -38,13 +38,11 @@ library
     Squeal.PostgreSQL.Expression.Composite
     Squeal.PostgreSQL.Expression.Json
     Squeal.PostgreSQL.Expression.Inline
-    Squeal.PostgreSQL.Expression.Inline2
     Squeal.PostgreSQL.Expression.Logic
     Squeal.PostgreSQL.Expression.Math
     Squeal.PostgreSQL.Expression.Null
     Squeal.PostgreSQL.Expression.Parameter
     Squeal.PostgreSQL.Expression.Range
-    Squeal.PostgreSQL.Expression.Range2
     Squeal.PostgreSQL.Expression.Set
     Squeal.PostgreSQL.Expression.Sort
     Squeal.PostgreSQL.Expression.Subquery
@@ -56,13 +54,10 @@ library
     Squeal.PostgreSQL.List
     Squeal.PostgreSQL.Manipulation
     Squeal.PostgreSQL.PG
-    Squeal.PostgreSQL.PG2
     Squeal.PostgreSQL.PQ
     Squeal.PostgreSQL.PQ.Connection
     Squeal.PostgreSQL.PQ.Decode
-    Squeal.PostgreSQL.PQ.Decode2
     Squeal.PostgreSQL.PQ.Encode
-    Squeal.PostgreSQL.PQ.Encode2
     Squeal.PostgreSQL.PQ.Exception
     Squeal.PostgreSQL.PQ.Indexed
     Squeal.PostgreSQL.PQ.Migration
