+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression.hs
@@ -356,7 +356,7 @@ function
   -> FunctionDB db x y
 function = unsafeFunction . renderSQL
 
--- | >>> printSQL $ unsafeFunctionN "f" (currentTime :* localTimestamp :* false *: inline 'a')
+-- | >>> printSQL $ unsafeFunctionN "f" (currentTime :* localTimestamp :* false *: inPG 'a')
 -- f(CURRENT_TIME, LOCALTIMESTAMP, FALSE, (E'a' :: char(1)))
 unsafeFunctionN :: SListI xs => ByteString -> xs ---> y
 unsafeFunctionN fun xs = UnsafeExpression $
